requirements.txt
python-telegram-bot==20.7
yt-dlp==2023.12.30
instaloader==4.10.3
requests==2.31.0
python-dotenv==1.0.0
asyncio==3.4.3
aiohttp==3.9.1
ffmpeg-python==0.2.0
Pillow==10.1.0
psutil==5.9.6

